+++ src/adapters/postgresql/schemas/index.ts
@@ -379,6 +379,8 @@ export {
   PartmanRunMaintenanceSchemaBase,
   PartmanShowPartitionsSchema,
   PartmanShowPartitionsSchemaBase,
+  PartmanShowConfigSchema,
+  PartmanShowConfigSchemaBase,
   PartmanCheckDefaultSchema,
   PartmanCheckDefaultSchemaBase,
   PartmanPartitionDataSchema,